Resolved conflicts possible problems with Menu Viewing with the refactor of circular list. Please check new circularlsit
Merge origin/master



Conflicts:
	src/CoreFunctions/MaintainMenu.java
	src/CoreFunctions/MenuDisplay.java

+++ src/entity/Items.java
@@ -12,13 +12,13 @@ package entity;
 public class Items {
     String ItemName;
     String Category;
-    int itemPrice;
+    double itemPrice;
     int quantity;
     public Items(){
         
     }
     
-    public Items(String ItemName, String Category, int itemPrice , int quantity){
+    public Items(String ItemName, String Category, double itemPrice , int quantity){
         this.ItemName = ItemName;
         this.Category = Category;
         this.itemPrice = itemPrice;
@@ -32,14 +32,14 @@ public class Items {
     public String getCategory(){
         return Category;
     }
-    public int getitemPrice(){
+    public double getitemPrice(){
         return itemPrice;
     }
     public int getQuantity(){
         return quantity;
     }
      public String toString() {
-        return String.format("Item Name : %-20s Category : %-20s  Price : %10d  Quantity : %10d\n",ItemName, Category , itemPrice , quantity);
+        return String.format("Item Name : %-20s Category : %-20s  Price : %10.2f  Quantity : %10d\n",ItemName, Category , itemPrice , quantity);
      }
